Move atomic file functions into atomic.go
They build a logical unit and separating those from the other file
functions makes that clear.
