+++ internal/pkg/file/file.go
@@ -20,9 +20,7 @@ import (
 	"errors"
 	"fmt"
 	"io"
-	"io/fs"
 	"os"
-	"path/filepath"
 
 	"github.com/k0sproject/k0s/internal/pkg/users"
 )
@@ -71,99 +69,6 @@ func Copy(src, dst string) (err error) {
 
 	return WriteAtomically(dst, sourceFileStat.Mode(), func(out io.Writer) error {
 		_, err = io.Copy(out, in)
-		return err
-	})
-}
-
-// WriteAtomically will atomically create or replace a file. The contents of the
-// file will be those that the write callback writes to the Writer that gets
-// passed in. The Writer will be unbuffered. WriteAtomically will buffer the
-// contents in a hidden (i.e. its name will start with a dot), temporary file
-// (it will have a .tmp extension). When write returns without an error, the
-// temporary file will be renamed to fileName, otherwise it will be deleted
-// without touching the target file.
-//
-// Note that this function is only best-effort on Windows:
-// https://github.com/golang/go/issues/22397#issuecomment-498856679
-func WriteAtomically(fileName string, perm os.FileMode, write func(file io.Writer) error) (err error) {
-	var fd *os.File
-	fd, err = os.CreateTemp(filepath.Dir(fileName), fmt.Sprintf(".%s.*.tmp", filepath.Base(fileName)))
-	if err != nil {
-		return err
-	}
-
-	tmpFileName := fd.Name()
-	close := true
-	defer func() {
-		var errs []error
-		if err != nil {
-			errs = append(errs, err)
-		}
-
-		if close {
-			if err := fd.Close(); err != nil {
-				errs = append(errs, err)
-			}
-		}
-
-		if err != nil {
-			err := os.Remove(tmpFileName)
-			// Don't propagate any fs.ErrNotExist errors. There is no point in
-			// doing this, since the desired state is already reached: The
-			// temporary file is no longer present on the file system.
-			if err != nil && !errors.Is(err, fs.ErrNotExist) {
-				errs = append(errs, err)
-			}
-		}
-
-		if len(errs) == 1 {
-			err = errs[0]
-		} else {
-			err = errors.Join(errs...)
-		}
-	}()
-
-	err = write(fd)
-	if err != nil {
-		return err
-	}
-
-	// https://github.com/google/renameio/blob/v2.0.0/tempfile.go#L150-L157
-	err = fd.Sync()
-	if err != nil {
-		return err
-	}
-
-	err = fd.Close()
-	close = false
-	if err != nil {
-		return err
-	}
-
-	err = os.Chmod(tmpFileName, perm)
-	if err != nil {
-		return err
-	}
-
-	err = os.Rename(tmpFileName, fileName)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// WriteContentAtomically will atomically create or replace a file with the
-// given content. WriteContentAtomically will create a hidden (i.e. its name
-// will start with a dot), temporary file (it will have a .tmp extension) with
-// the given content. Afterwards, the temporary file will be renamed to
-// fileName, otherwise it will be deleted without touching the target file.
-//
-// Note that this function is only best-effort on Windows:
-// https://github.com/golang/go/issues/22397#issuecomment-498856679
-func WriteContentAtomically(fileName string, content []byte, perm os.FileMode) error {
-	return WriteAtomically(fileName, perm, func(file io.Writer) error {
-		_, err := file.Write(content)
 		return err
 	})
 }
